+++ pkg/block/indexheader/reader_pool_test.go
@@ -22,8 +22,9 @@ import (
 
 func TestReaderPool_NewBinaryReader(t *testing.T) {
 	tests := map[string]struct {
-		lazyReaderEnabled     bool
-		lazyReaderIdleTimeout time.Duration
+		lazyReaderEnabled           bool
+		lazyReaderIdleTimeout       time.Duration
+		lazyReaderIdleDeleteTimeout time.Duration
 	}{
 		"lazy reader is disabled": {
 			lazyReaderEnabled: false,
@@ -59,7 +60,7 @@ func TestReaderPool_NewBinaryReader(t *testing.T) {
 
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
-			pool := NewReaderPool(log.NewNopLogger(), testData.lazyReaderEnabled, testData.lazyReaderIdleTimeout, NewReaderPoolMetrics(nil), AlwaysEagerDownloadIndexHeader)
+			pool := NewReaderPool(log.NewNopLogger(), testData.lazyReaderEnabled, testData.lazyReaderIdleTimeout, testData.lazyReaderIdleDeleteTimeout, NewReaderPoolMetrics(nil), AlwaysEagerDownloadIndexHeader)
 			defer pool.Close()
 
 			r, err := pool.NewBinaryReader(ctx, log.NewNopLogger(), bkt, tmpDir, blockID, 3, meta)
@@ -96,7 +97,7 @@ func TestReaderPool_ShouldCloseIdleLazyReaders(t *testing.T) {
 	testutil.Ok(t, err)
 
 	metrics := NewReaderPoolMetrics(nil)
-	pool := NewReaderPool(log.NewNopLogger(), true, idleTimeout, metrics, AlwaysEagerDownloadIndexHeader)
+	pool := NewReaderPool(log.NewNopLogger(), true, idleTimeout, 0, metrics, AlwaysEagerDownloadIndexHeader)
 	defer pool.Close()
 
 	r, err := pool.NewBinaryReader(ctx, log.NewNopLogger(), bkt, tmpDir, blockID, 3, meta)
@@ -131,4 +132,67 @@ func TestReaderPool_ShouldCloseIdleLazyReaders(t *testing.T) {
 	testutil.Assert(t, !pool.isTracking(r.(*LazyBinaryReader)))
 	testutil.Equals(t, float64(2), promtestutil.ToFloat64(metrics.lazyReader.loadCount))
 	testutil.Equals(t, float64(2), promtestutil.ToFloat64(metrics.lazyReader.unloadCount))
+}
+
+func TestReaderPool_ShouldDeleteIdleLazyReaders(t *testing.T) {
+	const idleTimeout = time.Second
+
+	ctx := context.Background()
+
+	tmpDir := t.TempDir()
+
+	bkt, err := filesystem.NewBucket(filepath.Join(tmpDir, "bkt"))
+	testutil.Ok(t, err)
+	defer func() { testutil.Ok(t, bkt.Close()) }()
+
+	// Create block.
+	blockID, err := e2eutil.CreateBlock(ctx, tmpDir, []labels.Labels{
+		{{Name: "a", Value: "1"}},
+		{{Name: "a", Value: "2"}},
+	}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "1"}}, 124, metadata.NoneFunc)
+	testutil.Ok(t, err)
+	testutil.Ok(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), metadata.NoneFunc))
+	meta, err := metadata.ReadFromDir(filepath.Join(tmpDir, blockID.String()))
+	testutil.Ok(t, err)
+
+	metrics := NewReaderPoolMetrics(nil)
+	pool := NewReaderPool(log.NewNopLogger(), true, idleTimeout, idleTimeout, metrics, AlwaysEagerDownloadIndexHeader)
+	defer pool.Close()
+
+	r, err := pool.NewBinaryReader(ctx, log.NewNopLogger(), bkt, tmpDir, blockID, 3, meta)
+	testutil.Ok(t, err)
+	defer func() { testutil.Ok(t, r.Close()) }()
+
+	// Ensure it can read data.
+	labelNames, err := r.LabelNames()
+	testutil.Ok(t, err)
+	testutil.Equals(t, []string{"a"}, labelNames)
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.loadCount))
+	testutil.Equals(t, float64(0), promtestutil.ToFloat64(metrics.lazyReader.unloadCount))
+	testutil.Equals(t, float64(0), promtestutil.ToFloat64(metrics.lazyReader.deleteCount))
+
+	// Wait enough time before checking it.
+	time.Sleep(idleTimeout * 2)
+
+	// We expect the reader has been closed, but not released from the pool.
+	testutil.Assert(t, pool.isTracking(r.(*LazyBinaryReader)))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.loadCount))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.unloadCount))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.deleteCount))
+
+	// Ensure it can still read data (will be re-opened).
+	labelNames, err = r.LabelNames()
+	testutil.Ok(t, err)
+	testutil.Equals(t, []string{"a"}, labelNames)
+	testutil.Assert(t, pool.isTracking(r.(*LazyBinaryReader)))
+	testutil.Equals(t, float64(2), promtestutil.ToFloat64(metrics.lazyReader.loadCount))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.unloadCount))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.deleteCount))
+
+	// We expect an explicit call to Close() to close the reader and release it from the pool too.
+	testutil.Ok(t, r.Close())
+	testutil.Assert(t, !pool.isTracking(r.(*LazyBinaryReader)))
+	testutil.Equals(t, float64(2), promtestutil.ToFloat64(metrics.lazyReader.loadCount))
+	testutil.Equals(t, float64(2), promtestutil.ToFloat64(metrics.lazyReader.unloadCount))
+	testutil.Equals(t, float64(1), promtestutil.ToFloat64(metrics.lazyReader.deleteCount))
 }